+++ CHANGELOG.md
@@ -4,8 +4,9 @@
 
 ## Added
 - Added the `nginx` command to with `restart`, `stop`, and `start` commands. 
-- Added the `nitrod` daemon that runs in the virtual machine and exposes an API on port `9999`. New Nitro features will be implemented using the API.
-- Added the `php` command to with `restart`, `stop`, and `start` commands. 
+- Added the `nitrod` daemon that runs in the virtual machine and exposes a gRPC API on port `50051`. New Nitro features will be implemented using the API.
+- Added the `php` command to with `restart`, `stop`, and `start` commands. ([#57](https://github.com/craftcms/nitro/issues/57))
+- Added the `iniset` command under `php` to quickly modify init settings for `max_execution_time`, `max_input_vars`, `max_input_time`, `upload_max_filesize`, `max_file_uploads`, and `memory_limit`. ([#138](https://github.com/craftcms/nitro/issues/138))
 
 ### 1.0.0-beta.10 - 2020-06-04